Propose an API for expiring scalars
Redis' feature for expiring keys has surfaced in [counters][] and in
[flags][]. I've found myself wanting expiring keys for scalars and
lists, but forcing myself into using flags, and writing manual
expiration code.

So this first commit is the beginning of an exploration of what it could
look like to make _everything_ expirable. I'm starting with scalars as a
proof of concept.

For this first attempt, I want to be able to leverage two Redis
commands:

- [EXPIRE][] for expiring a key in a relative amount of time
- [EXPIREAT][] for expiring a key at a specific point in time

[counters]: https://github.com/rails/kredis/blob/5d576221a9aadc6c73aa67229864bce623ac3d54/test/types/counter_test.rb#L39
[flags]: https://github.com/rails/kredis/blob/1cb2f295fdf622cf44e6d906b28a2c61bb7ab326/test/types/flag_test.rb#L16
[expire]: https://redis.io/commands/expire
[expireat]: https://redis.io/commands/expireat