Add Kredis::ScriptEvaluator to handle efficient script evals
The ScriptEvaluator first attempts an EVALSHA command with the SHA1
digest of the provided script, and if it errors, it will fallback to a
normal EVAL command.

Additionally it caches the script digests for future EVALs. So scripts
passed to the ScriptEvaluator should be frozen so that the digest cache
does not balloon.