Improved version of UniqueList: OrderedSet (#76)
* Back UniqueList with a Redis Sorted Set

A sorted set is optimal for this type of functionality because we can
use fewer, more-performant Redis calls.

* Wrap unique list insert and trim in a transaction

* Fallback / migration mechanism for existing users of UniqueLists

In order to ease the transition from UniqueList being backed by
a Redis list, we can fallback to the legacy implementation for read
operations. For write operations we first migrate to a sorted set,
then retry.

* Update the Readme with sorted set backed unique list commands

* Test remaining methods with migration fallbacks

* Be more specific about which errors to raise

* Improve the fallback / migration system

* Improve unique list scoring to avoid collisions

Use the Redis server time as a basis. Then add the current process time.
Then add to that the index of the inserted elements as a microsecond
component.

* Update Readme with realistic Unique List scores

* Implement OrderedSet

Like a UniqueList, but backed by a Redis sorted set rather than a list

* Resolve pipeline deprecation warnings

* Document OrderedSet

* Remove UniqueListLegacy

* Revert fallback options

* Revert change to type_from

* Enforce limit is not negative

* Revert config proxying

* Fix bug in score ge
neration