Global namespacing that applies to all threads (#158)
Namespacing is thread-local currently, so it's easy to lose, e.g.
when running parallel tests, Capybara tests, or executor pools.

Changing that is a major compatibility break, so we introduce a
non-thread-local Kredis.global_namespace attribute that behaves
as we'd expect.

When Kredis.namespace is set, it's appended to the global namespace.